+++ src/index.tsx
@@ -1,2 +1,3 @@
 export { default as CreditCard } from "./components/CreditCard";
+export * from "./utils/get-card-issuer";
 export type Focused = "number" | "name" | "expiry" | "cvc" | "";