+++ src/lib/intl.ts
@@ -3,12 +3,15 @@ export type Locale = "en" | "pt-BR" | "es";
 const messages = {
   en: {
     validThru: "VALID THRU",
+    namePlaceholder: "YOUR NAME HERE",
   },
-  'pt-BR': {
+  "pt-BR": {
     validThru: "VÁLIDO ATÉ",
+    namePlaceholder: "SEU NOME AQUI",
   },
   es: {
     validThru: "VÁLIDO HASTA",
+    namePlaceholder: "SU NOMBRE AQUÍ",
   },
 };